Extracted test data and added alias to tsconfig
